j'ai traviei sur tach de admin

+++ app/Http/Controllers/CategorieController.php
@@ -4,8 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Models\Categorie;
 use App\Http\Controllers\Controller;
+use App\Models\User;
 use Illuminate\Http\Request;
-
+use Illuminate\Support\Facades\Auth;
 
 class CategorieController extends Controller
 {
@@ -31,43 +32,63 @@ class CategorieController extends Controller
     public function store(Request $request)
     {
 
-        // dd($request);
-        $categorie =  $request->validate([
-            'name' => 'required|string|max:255|unique:categories',
-        ]);
-
-        if ($categorie) {
-            Categorie::create([
-                'name' => $categorie['name'],
-                'user_id' => 1,
+        if (Auth::check()) {
+            $user = Auth::user();
+            $categorie =  $request->validate([
+                'name' => 'required|string|max:255|unique:categories',
             ]);
 
-            return redirect()->route('profile')->with('status', 'Categorie créé avec succès');
+            if ($user->roleId->name != 'admin') {
+
+                return redirect()->route('home');
+            }
+
+            if ($categorie) {
+                Categorie::create([
+                    'name' => $categorie['name'],
+                    'user_id' => 1,
+                ]);
+
+                return redirect()->route('admin')->with('status', 'Categorie créé avec succès');
+            }
+        } else {
+            return redirect()->route('home');
         }
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Categorie $categorie)
-    {
-        //
-    }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Categorie $categorie)
-    {
-        //
-    }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Categorie $categorie)
+    public function update(Request $request)
     {
-        //
+        if (Auth::check()) {
+            $user = Auth::user();
+
+            if ($user->roleId->name != 'admin') {
+                return redirect()->route('home')->with('status', 'Unauthorized');
+            }
+
+            $categorie = Categorie::find($request->id);
+
+            if (!$categorie) {
+                return redirect()->route('home')->with('status', 'Category not found');
+            }
+
+            $validatedData = $request->validate([
+                'name' => 'required|string|max:255',
+            ]);
+
+            $categorie->update($validatedData);
+
+            return redirect()->route('admin')->with('status', 'Category updated successfully');
+        } else {
+            return redirect()->route('home')->with('status', 'Unauthorized');
+        }
     }
 
     /**
@@ -75,16 +96,24 @@ class CategorieController extends Controller
      */
     public function destroy(Request $request)
     {
-        $categorie = Categorie::find($request->id);
+        if (Auth::check()) {
+            $user = Auth::user();
 
+            if ($user->roleId->name != 'admin') {
+                return redirect()->route('home')->with('status', 'Unauthorized');
+            }
 
-        if (!$categorie) {
+            $categorie = Categorie::find($request->id);
 
-            return redirect()->route('profile')->with('status', 'Categorie not found');
+            if (!$categorie) {
+                return redirect()->route('home')->with('status', 'Category not found');
+            }
+
+            $categorie->delete();
+
+            return redirect()->route('admin')->with('status', 'Category deleted successfully');
+        } else {
+            return redirect()->route('home')->with('status', 'Unauthorized');
         }
-
-        $categorie->delete();
-
-        return redirect()->route('profile')->with('status', 'Categorie est suppremer');
     }
 }
