+++ app/Http/Controllers/EvenmentController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
+use App\Models\Categorie;
 use App\Models\Evenment;
 use App\Models\Reservation;
 use App\Models\User;
@@ -16,12 +17,12 @@ class EvenmentController extends Controller
      */
     public function index()
     {
-
-        // $evenement = Evenment::with('user')->with('ville')->get();
-        $evenement = Evenment::with('user')->with('ville')->with('catecory')->latest()->paginate(6);
-
-        // dd($evenement);
-        // return response()->json($evenement);
+        $evenement = Evenment::where('accepter', 1)->latest()->paginate(6);
+        $user = Auth::user();
+        if($user){
+            $reservation = Reservation::where('user_id' , $user->id)->get();
+            return view('home', compact('evenement' , 'reservation'));
+        }
 
         return view('home', compact('evenement'));
     }
@@ -39,9 +40,16 @@ class EvenmentController extends Controller
             ->with('ville')
             ->with('catecory')
             ->where('titre', 'like', '%' . $query . '%')
+            ->where('accepter', 1)
             ->latest()
             ->paginate(6);
 
+        // $evenement = Categorie::where('name', 'like', '%' . $query . '%')
+        // ->with('even')
+        // ->latest()
+        // ->paginate(6);
+
+        // return  response()->json($evenement);
 
         return view('home', compact('evenement'));
     }
@@ -52,6 +60,10 @@ class EvenmentController extends Controller
     public function store(Request $request)
     {
         $user = Auth::user();
+
+        if ($user->roleId->name == 'utilisateur') {
+            return redirect()->route('home')->with('status', 'Unauthorized action');
+        }
 
         $evenement = $request->validate([
             'titre' => 'required|string|min:3|max:30',
@@ -67,6 +79,9 @@ class EvenmentController extends Controller
             'ville_id' => $request->ville,
             'nombre_places' => $evenement["nombre_places"],
             'categorie_id' => $request->categorie_id,
+            'prix' => $request->prix,
+            'auto' => $request->auto,
+            'accepter' => 0,
             'user_id' => $user->id,
         ]);
 
@@ -78,11 +93,80 @@ class EvenmentController extends Controller
      */
     public function show(Request $request)
     {
+        
 
         $evenement = Evenment::where('id', $request->id)->first();
 
+       
+
+
+        if ($evenement->accepeter != 0) {
+
+            return back()->with('status', 'Désolé, les réservations pour cet événement ne sont pas encore ouvertes');
+        }
+
         return view('detalise', compact("evenement"));
     }
+
+    public function show1(Request $request)
+    {
+        $user = Auth::user();
+
+
+
+        $request->validate([
+            'id' => 'required|exists:evenments,id',
+        ]);
+
+        $evenement = Evenment::findOrFail($request->id);
+
+
+        if ($evenement->accepeter == 0) {
+
+            return back()->with('status', 'Désolé, les réservations pour cet événement ne sont pas encore ouvertes');
+        }
+
+        if ($user->id != $evenement->user_id) {
+
+            if ($evenement->accepeter == 0) {
+                return back()->with('status', 'Désolé, les réservations pour cet événement ne sont pas encore ouvertes');
+            }
+
+            return view('detalise', compact('evenement'));
+        }
+
+
+
+        return view('detalise', compact("evenement"));
+    }
+
+
+    public function accepter(Request $request)
+    {
+        if (Auth::check()) {
+            $user = Auth::user();
+
+            $evenement = Evenment::where('id', $request->id)->first();
+
+            if (!$evenement) {
+                return redirect()->route('home')->with('status', 'Event not found or unauthorized action');
+            }
+
+            if ($user->roleId->name != 'admin') {
+                return redirect()->route('home')->with('status', 'Unauthorized action');
+            }
+
+            $evenement->accepter = true;
+            $evenement->save();
+
+            return back()->with('status', 'Event accepted successfully');
+        }
+
+        return redirect()->route('login');
+    }
+
+
+
 
     /**
      * Show the form for editing the specified resource.
@@ -90,8 +174,6 @@ class EvenmentController extends Controller
     public function edit(Request $request)
     {
         $evenement = Evenment::where('id', $request->id)->first();
-
-        // dd($evenement);
         return view('edite', compact("evenement"));
     }
 
@@ -100,35 +182,56 @@ class EvenmentController extends Controller
      */
     public function update(Request $request)
     {
+        if (Auth::check()) {
+            $user = Auth::user();
+            $evenement = Evenment::find($request->id);
 
-        $evenement = Evenment::find($request->id);
+            if (!$evenement) {
+                return redirect()->route('profile')->with('status', 'Event not found');
+            }
 
+            if ($evenement->user_id !== $user->id) {
 
-        if (!$evenement) {
-            return response()->json(['evenement' => 'Evenment not found'], 404);
+                return redirect()->route('profile')->with('status', 'Unauthorized');
+            }
+
+            $validatedData = $request->validate([
+                'titre' => 'required|string',
+                'description' => 'required|string'
+            ]);
+
+            $evenement->update($validatedData);
+
+            return redirect()->route('profile')->with('status', 'Event updated successfully');
+        } else {
+            return redirect()->route('profile')->with('status', 'Unauthorized');
         }
-
-        $evenement->update($request->all());
-
-        return redirect()->route('profile');
     }
+
+
 
     /**
      * Remove the specified resource from storage.
      */
     public function destroy(Request $request)
     {
-        $evenement = Evenment::find($request->id);
+        if (Auth::check()) {
+            $user = Auth::user();
+            $evenement = Evenment::find($request->id);
 
+            if (!$evenement) {
+                return redirect()->route('profile')->with('status', 'Event not found');
+            }
 
-        if (!$evenement) {
+            if ($evenement->user_id !== $user->id) {
+                return redirect()->route('profile')->with('status', 'Unauthorized');
+            }
 
-            return redirect()->route('profile')->with('status', 'Evenment not found');
-            
+            $evenement->delete();
+
+            return redirect()->route('profile')->with('status', 'Event deleted successfully');
+        } else {
+            return redirect()->route('profile')->with('status', 'Unauthorized');
         }
-
-        $evenement->delete();
-
-        return redirect()->route('profile')->with('status', 'Evenment est suppremer');
     }
 }
